+++ src/main.jsx
@@ -5,7 +5,9 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 //Här importerar vi alla sidor som kommer vilja använda i routingen
 import App from './App.jsx';
 import Home from './Home.jsx';
-import Movies from './Movies.jsx';
+import Movies from './components/Filmer/Movies.jsx';
+import MovieItem from './components/Filmer/MovieItem.jsx';
+import MovieFilter from './components/Filmer/MovieFilter.jsx';
 import ContactUs from './ContactUs.jsx';
 import Booking from './Booking.jsx';
 import BecomeMember from './BecomeMember.jsx';
@@ -24,16 +26,16 @@ import CancelBooking from './CancelBooking.jsx'
 //den översta i pages ska vara startsidan
 
 export const pages = [
-  { path: '/', label: 'Hem', element: <Home /> },
-  { path: '/Filmer', label: 'Filmer', element: <Movies /> },
-  { path: '/boka', label: 'Boka', element: <Booking /> },
-  { path: '/kontakt', label: 'Kontakta oss', element: <ContactUs /> },
-  { path: '/blimedlem', label: 'Bli medlem', element: <BecomeMember /> },
-  { path: '/loggain', label: 'Logga in', element: <LogIn /> },
-  { path: '/detaljsidan/:movieId', label: '', element: <MoviePage /> },
-  { path: '/finalize-booking', label: '', element: <FinalizeBooking /> }, // Ta bort sedan, kunder/externa användare ska inte kunna se denna sida utan att ha valt film.
+    { path: '/', label: 'Hem', element: <Home /> },
+    { path: '/Filmer', label: 'Filmer', element: <Movies /> },
+    { path: '/boka', label: 'Boka', element: <Booking /> },
+    { path: '/kontakt', label: 'Kontakta oss', element: <ContactUs /> },
+    { path: '/blimedlem', label: 'Bli medlem', element: <BecomeMember /> },
+    { path: '/loggain', label: 'Logga in', element: <LogIn /> },
+    { path: '/detaljsidan/:movieId', label: '', element: <MoviePage /> },
+    { path: '/finalize-booking', label: '', element: <FinalizeBooking /> }, // Ta bort sedan, kunder/externa användare ska inte kunna se denna sida utan att ha valt film.
     { path: '/bokningsbekraftelse', label: '', element: <BookingConfirmation /> }, // Ta bort sedan, kunder/externa användare ska inte kunna se denna sida utan att ha valt film.
-    { path: '/avbokning', label: '', element: <CancelBooking /> } 
+    { path: '/avbokning', label: '', element: <CancelBooking /> }
 ];
 
 const router = createBrowserRouter([
