+++ src/sass/movies.css
@@ -1,0 +1,57 @@
+/*Movies.css */
+.movie-poster {
+    width: 100%;
+    height: 400px;
+    object-fit: cover;
+    border: 2px solid black;
+}
+
+/*movie-filter.css */
+
+.age-filter {
+    margin-bottom: 15px;
+}
+
+.genre-filter {
+    margin-bottom: 15px;
+}
+
+.age-filter .btn {
+    background-color: #ff5722;
+    /* Byt ut med den färg du föredrar */
+}
+
+.genre-filter .btn {
+    background-color: #2196f3;
+    /* Byt ut med den färg du föredrar */
+}
+
+/*Movie-item*/
+
+.css .movie-item {
+    margin-bottom: 10px;
+}
+
+.poster img {
+    width: 150px;
+    height: auto;
+    border: 2px solid black;
+}
+
+.movie-info {
+    display: inline-block;
+    margin-left: auto;
+    /* Justera avståndet mellan postern och info-raden */
+}
+
+.movie-title {
+    font-size: 20px;
+    margin-bottom: 5px;
+    color: white;
+}
+
+.movie-description {
+    margin: 0;
+    font-size: 16px;
+    color: white;
+}
